requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.9
    # via microsoft-kiota-authentication-azure
aiosignal==1.3.1
    # via aiohttp
anyio==4.6.0
    # via httpx
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
azure-core==1.31.0
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.18.0
    # via
    #   -r requirements.in
    #   msgraph-sdk
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==43.0.1
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
exceptiongroup==1.2.2
    # via anyio
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.6
    # via httpx
httpx[http2]==0.27.2
    # via
    #   httpx
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via opentelemetry-api
microsoft-kiota-abstractions==1.3.3
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-authentication-azure==1.1.0
    # via
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-http==1.3.3
    # via
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-serialization-form==0.1.1
    # via msgraph-sdk
microsoft-kiota-serialization-json==1.3.3
    # via msgraph-sdk
microsoft-kiota-serialization-multipart==0.1.0
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.0.0
    # via msgraph-sdk
msal==1.31.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
msgraph-core==1.1.5
    # via msgraph-sdk
msgraph-sdk==1.9.0
    # via -r requirements.in
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.27.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.27.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
pendulum==3.0.0
    # via
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
portalocker==2.10.1
    # via msal-extensions
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via
    #   msal
    #   pyjwt
python-dateutil==2.9.0.post0
    # via
    #   microsoft-kiota-serialization-text
    #   pendulum
    #   time-machine
requests==2.32.3
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
std-uritemplate==2.0.0
    # via microsoft-kiota-abstractions
time-machine==2.15.0
    # via pendulum
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-identity
    #   multidict
    #   opentelemetry-sdk
tzdata==2024.2
    # via pendulum
urllib3==2.2.3
    # via requests
wrapt==1.16.0
    # via deprecated
yarl==1.13.1
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata
